+++ bakedanuki/bakedanuki-util/python/bd_util/maya/node/operator/attr/define/custom/at/scalar_compound/numeric_compound/double_compound/double4_compound/_base.py
@@ -1,8 +1,10 @@
 # coding: utf-8
-from typing import Any, TypeVar, Type, cast
+from typing import Any, Generic, TypeVar, Type, cast
 
 # self
-from ...........value import Double4
+from ...........value.scalar_compound.scalar_compound_value import (
+    ScalarCompoundValue,
+)
 from ............ import logger as u_logger
 from .._base import (
     DoubleCompoundBasePlugOperator,
@@ -12,18 +14,19 @@ from .._base import (
 
 A = TypeVar("A", bound="Double4CompoundBaseAttrOperator[Any]")
 
-P = TypeVar("P", bound="Double4CompoundBasePlugOperator[Any]")
+P = TypeVar("P", bound="Double4CompoundBasePlugOperator[Any, Any]")
+
+V = TypeVar("V", bound=ScalarCompoundValue[float])
 
 
 logger = u_logger.get_logger(__name__, level=u_logger.DEBUG)
 
 
 class Double4CompoundBasePlugOperator(
-    DoubleCompoundBasePlugOperator[A, Double4]
+    DoubleCompoundBasePlugOperator[A, V],
+    Generic[A, V],
 ):
     __slots__ = ()
-
-    VALUE_TYPE = Double4
 
 
 class Double4CompoundBaseAttrOperator(DoubleCompoundBaseAttrOperator[P]):