+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/__init__.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from .double import Double2, Double3, Double4, Quat
+from .double import Double2, Double3, Double4
 from .float import Float2, Float3
 from .long import Long2, Long3
 from .short import Short2, Short3
@@ -12,7 +12,6 @@ __all__ = (
     "Float3",
     "Long2",
     "Long3",
-    "Quat",
     "Short2",
     "Short3",
 )