+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/double/double2/double2.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 from ....scalar2 import Scalar2
+from ..._floating_arithmetic import FloatingNumericArithmeticMixin
 
 
-class Double2(Scalar2[float]):
+class Double2(FloatingNumericArithmeticMixin, Scalar2[float]):
     __slots__ = ()