+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/double/double3/double3.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 from ....scalar3 import Scalar3
+from ..._floating_arithmetic import FloatingNumericArithmeticMixin
 
 
-class Double3(Scalar3[float]):
+class Double3(FloatingNumericArithmeticMixin, Scalar3[float]):
     __slots__ = ()