+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/double/double4/__init__.py
@@ -1,5 +1,4 @@
 # coding: utf-8
 from .double4 import Double4
-from .quat import Quat
 
-__all__ = ("Double4", "Quat")
+__all__ = ("Double4",)