+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/double/double4/double4.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 from ....scalar4 import Scalar4
+from ..._floating_arithmetic import FloatingNumericArithmeticMixin
 
 
-class Double4(Scalar4[float]):
+class Double4(FloatingNumericArithmeticMixin, Scalar4[float]):
     __slots__ = ()