+++ bakedanuki/bakedanuki-util/python/bd_util/maya/value/scalar_compound/numeric/float/float3/float3.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 from ....scalar3 import Scalar3
+from ..._floating_arithmetic import FloatingNumericArithmeticMixin
 
 
-class Float3(Scalar3[float]):
+class Float3(FloatingNumericArithmeticMixin, Scalar3[float]):
     __slots__ = ()